refactor: combine multiple params into a uuid one
A lot of different routes need a parameter which is just the uuid.
This is always needed to validate it as an uuid. Instead of having them
redundantly, just use one dto for it. In case we need to differentiate
later one, we can just make a partial type and reuse that one instead.